etl-pandas/requirements.txt
pandas
sqlalchemy
pyyaml
oracledb
